fixes #10
The main reason of this problem is that,
using word from user input as defaut Record.word
and use it to check if this word have already in db or not.
If db has it, just use the result store in db.
If not, send a query for this word.
Then update it if bs4 got <span class="yschttl"> string

So, no matter "iblis", "Iblis", "IBLIS", "iBlis", etc...
all of these words are regarded as new words at the first time
but come up with the totally same result Iblis.
Then, the program will save the Record.word only as Iblis
which lead to the Unique Key error.

To fix this bug,
make all the Record.word as lowercase for Record.get().
Then, add a new content['word'] for storing the word from the query result.