Add wiktionary (#269)
**_This Pull Request refers to #196 issue._**

### Description:

**The API problem.** 
We can access wiktionary content through [WIKIMEDIA REST API](https://en.wiktionary.org/api/rest_v1/). Unfortunately this does not support ZH language (at least I think so). The [main API](https://en.wiktionary.org/w/api.php) supports it but it requires a ton of work to extract anything from it (Reason: The wiktionary is designed to be human readable). There are APIs that intend to provide wiktionary data but could not find anything that would not require to implement an API key.

**Support for more languages**
The WIKIMEDIA REST API does provide definitions in other languages, so it is probably a good idea to implement a way to support them using CLI arguments in the future.